Disconnect after INACTIVE_TIMEOUT (60) seconds, releasing Postgres worker.
In order to take advantage of the LIFO queue in poolboy, we create
each database worker in a detached state.  Only when a request comes
in do we attempt to connect.  On each request that comes in, we reset
the timeout to INACTIVE_TIMEOUT seconds - 60 for the time being.

At present, this has a negative aspect to it: if a web site were to
sit idle, all its database workers would time out, meaning that the
first person to request a page after the idle period would get a
slightly slower page load because of having to reopen the database
connection and worker.